requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
selenium
webdriver-manager
ollama
python-dotenv
